+++ handlers/test.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"log/slog"
 	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
@@ -81,7 +80,7 @@ type IncomingTestCase struct {
 func testHandlerRequest(tb testing.TB, s *courier.Server, path string, headers map[string]string, data string, multipartFormFields map[string]string, expectedStatus int, expectedBodyContains string, requestPrepFunc RequestPrepFunc) string {
 	var req *http.Request
 	var err error
-	url := fmt.Sprintf("https://%s%s", s.Config().Domain, path)
+	url := fmt.Sprintf("https://%s%s", s.Runtime().Config.Domain, path)
 
 	if data != "" {
 		req, err = http.NewRequest(http.MethodPost, url, strings.NewReader(data))
@@ -140,7 +139,6 @@ func testHandlerRequest(tb testing.TB, s *courier.Server, path string, headers m
 
 func newServer(backend courier.Backend) *courier.Server {
 	// for benchmarks, log to null
-	logger := slog.Default()
 	log.SetOutput(io.Discard)
 
 	cfg := runtime.NewDefaultConfig()
@@ -148,8 +146,7 @@ func newServer(backend courier.Backend) *courier.Server {
 	cfg.FacebookApplicationSecret = "fb_app_secret"
 	cfg.WhatsappAdminSystemUserToken = "wac_admin_system_user_token"
 
-	return courier.NewServerWithLogger(cfg, backend, logger)
-
+	return courier.NewServer(runtime.NewTestRuntime(cfg), backend)
 }
 
 // RunIncomingTestCases runs all the passed in tests cases for the passed in channel configurations